requirements.txt
fastapi
uvicorn
pymupdf
chromadb
python-dotenv
python-multipart
groq
sentence-transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.11.0+cpu